requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1

numpy==1.26.4
pillow==10.3.0
tqdm==4.66.4
nltk

# BLIP (captioning)
transformers==4.43.2
torch>=2.2.0
sentencepiece==0.2.0
timm==0.9.12
huggingface_hub==0.24.5

# OpenCLIP (embeddings)
open_clip_torch==2.24.0

# Vector DB
qdrant-client==1.9.2

